backend/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.5
contourpy==1.1.1
cycler==0.11.0
dill==0.3.7
dnspython==2.4.2
filelock==3.12.4
fonttools==4.42.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
lit==17.0.1
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
psutil==5.9.5
py-cpuinfo==9.0.0
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.188
urllib3==2.0.5
